+++ checkpoint/CHANGELOG.md
@@ -9,8 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added an option `use_zarr2_compression` to `PyTreeCheckpointHandler` to
-  control whether compress will be used for zarr2. Default is kept `True`.
+- Added an option `use_compression` to `PyTreeCheckpointHandler` to
+  control whether compress will be used. Default is kept `True`.
 
 ### Changed